+++ gateway/src/main/java/org/acumos/federation/gateway/GatewayController.java
@@ -37,11 +37,14 @@ import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.client.RestClientResponseException;
 import org.springframework.security.access.annotation.Secured;
-
+import org.springframework.security.access.method.P;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.acumos.cds.domain.MLPCatalog;
 import org.acumos.cds.domain.MLPSolution;
 import org.acumos.cds.domain.MLPPeer;
@@ -50,6 +53,9 @@ import org.acumos.cds.domain.MLPPeerSubscription;
 import org.acumos.federation.client.FederationClient;
 import org.acumos.federation.client.GatewayClient;
 import org.acumos.federation.client.data.JsonResponse;
+import org.acumos.federation.client.data.ModelData;
+import org.acumos.federation.client.data.ModelInfo;
+
 
 
 /**
@@ -70,6 +76,9 @@ public class GatewayController {
 
 	@Autowired
 	private SubscriptionPoller poller;
+
+	@Autowired
+	private WebSecurityConfigurerAdapter security;
 
 	@ApiOperation(value = "Invoked by local Acumos to get a list of catalogs available from a peer Acumos instance .", response = MLPCatalog.class, responseContainer = "List")
 	@GetMapping(FederationClient.CATALOGS_URI)
@@ -151,6 +160,76 @@ public class GatewayController {
 			response.setStatus(HttpServletResponse.SC_OK);
 		}
 		return ret;
+	}
+	/**
+	 * Receives incoming log message from logstash and Sends to {@link FederationController#receiveModelData(ModelData, HttpServletResponse)}
+	 *
+	 * @param payload model data payload The payload must have a model.solutionId
+	 *
+	 * @param theHttpResponse HttpServletResponse
+	 * @param peerIdPathVar PeerID from url path param or USE_SOLUTION_SOURCE to lookup peer based on model.solutionId field
+	 * @return success message in JSON format
+	 *
+	 */
+	@Secured(Security.ROLE_PEER)
+	@ApiOperation(
+			value = "Invoked by local Acumos to post incoming model data to respective remote peer Acumos instance .",
+			response = ModelData.class)
+	@PostMapping(FederationClient.MODEL_DATA)
+	@ResponseBody
+	public JsonResponse<Void> peerModelData(HttpServletResponse theHttpResponse,
+	    @RequestBody ModelData payload, @PathVariable("peerId") String peerIdPathVar) {
+		log.debug("/peer/{}/modeldata  payload: {}", peerIdPathVar, payload);
+		ModelInfo modelInfo = payload.getModel();
+		String peerId = peerIdPathVar;
+		JsonResponse response = new JsonResponse();
+		// peer id lookup from solution if peerid from path variable is null
+		if(peerId.indexOf("USE_SOLUTION_SOURCE") != -1){
+			String solutionId = modelInfo.getSolutionId();
+			peerId = getPeerIdFromCds(solutionId);
+		}
+		
+		MLPPeer self = ((Security) security).getSelf();
+		modelInfo.setSubscriberName(self.getSubjectName());
+	
+		try {
+
+			// check if thePeerId matches to the
+			// Ignore request if for local peer i.e. peerId same as local peer
+			//
+			log.debug("Attempting to connect to peer id {}", peerId);
+			if (peerId == null) {
+				log.debug("ignore logging to self-peer {}", peerId);
+				return this.getSuccessResponse(theHttpResponse,
+						"ignore logging to self-peer");
+			}
+
+			log.debug("calling peer with request {}", payload);
+			callPeer(theHttpResponse, peerId, peer -> peer.receiveModelData(payload));
+		} catch (Exception ex) {
+			log.error("failed posting to remote peerId:" + peerId + " exception {}", ex);
+			throw new BadRequestException(HttpServletResponse.SC_BAD_GATEWAY, "failed posting to remote peerId:" + peerId);
+		}
+		return response;
+
+	}
+
+	private JsonResponse getSuccessResponse(
+		HttpServletResponse theHttpResponse,
+		String message) {
+		JsonResponse response = new JsonResponse();
+		response.setMessage("modelData - " + message);
+		return response;
+	}
+
+	private String getPeerIdFromCds(String solutionId) {
+		try {
+			String peerId = clients.getCDSClient().getSolution(solutionId).getSourceId();
+			return peerId;
+		} catch (RestClientResponseException ex) {
+			log.error("getSolution failed, server reports: {}", ex);
+			throw new BadRequestException(HttpServletResponse.SC_NOT_FOUND, "Not Found");
+		}
 	}
 
 	private <T> JsonResponse<T> callPeer(HttpServletResponse response, String peerId, Function<FederationClient, T> fcn) {