+++ gateway/src/main/java/org/acumos/federation/gateway/LogstashServiceImpl.java
@@ -1,0 +1,37 @@
+/*-
+ * ===============LICENSE_START=======================================================
+ * Acumos
+ * ===================================================================================
+ * Copyright (C) 2020 Nordix Foundation
+ * ===================================================================================
+ * This Acumos software file is distributed by Nordix Foundation
+ * under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * This file is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ===============LICENSE_END=========================================================
+ */
+package org.acumos.federation.gateway;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.client.RestClientException;
+import org.acumos.federation.client.data.ModelData;
+
+public class LogstashServiceImpl implements LogstashService {
+
+
+	@Autowired
+	private Clients clients;
+
+	@Override
+	public void sendModelData(ModelData payload) throws RestClientException{
+	 clients.getLogstashClient().saveModelData(payload);
+	}
+
+}